+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## alpha 2.1.0-alpha.1
 - don't reset the layer limit after parsing the gcode
+- highlight the top layer and the last segment
+- top layer / last segment color is configurable
 
 ## alpha 2.1.0-alpha.0
 - add incremental parsing